requirements.txt
aioredis==2.0.1
anyio==3.6.1
async-timeout==4.0.2
click==8.1.3
fastapi==0.79.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
pydantic==1.9.1
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.3.0
uuid==1.30
uvicorn==0.18.2
websockets==10.3
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.0
click==8.1.3
fastapi==0.79.0
frozenlist==1.3.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
multidict==6.0.2
pydantic==1.9.1
python-dotenv==0.20.0
redis==3.5.3
rejson==0.5.6
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing-extensions==4.3.0
uuid==1.30
uvicorn==0.18.2
websockets==10.3
yarl==1.7.2
